activy register fragment implement logic and saves to viewmodel

+++ app/src/main/java/study/project/models/User.kt
@@ -22,7 +22,7 @@ data class User(
     var email: String = "",
 
     @ColumnInfo(name = "weight")
-    var weight: Int = 0,
+    var weight: Double = 0.0,
 
     @ColumnInfo(name = "height")
     var height: Int = 0,
